Move config to new dir
