+++ src/client/actions.ts
@@ -24,7 +24,7 @@ import {
   StatusOr,
   Food,
 } from "../core";
-import { defaultConfig } from "./config";
+import { defaultConfig } from "../config/config";
 import { fetchFdcFood, maybeRewriteFoodReference } from "../food_data_central";
 
 export type ThunkResult<R> = ThunkAction<R, RootState, undefined, RootAction>;
